feat(solver): add penetration constraint plugin on solver plugin and change solver file name
