+++ engine/CMakeLists.txt
@@ -71,7 +71,10 @@ set(CUBOS_ENGINE_SOURCE
 
 	"src/physics/plugin.cpp"
 	"src/physics/gravity.cpp"
-	"src/physics/solver.cpp"
+	"src/physics/solver/plugin.cpp"
+	"src/physics/solver/utils.cpp"
+	"src/physics/constraints/penetration_constraint.cpp"
+	"src/physics/solver/penetration_constraint/plugin.cpp"
 
 	"src/input/plugin.cpp"
 	"src/input/input.cpp"
