+++ engine/src/defaults/plugin.cpp
@@ -6,7 +6,7 @@
 #include <cubos/engine/imgui/plugin.hpp>
 #include <cubos/engine/input/plugin.hpp>
 #include <cubos/engine/physics/plugin.hpp>
-#include <cubos/engine/physics/solver/solver.hpp>
+#include <cubos/engine/physics/solver/plugin.hpp>
 #include <cubos/engine/renderer/plugin.hpp>
 #include <cubos/engine/scene/plugin.hpp>
 #include <cubos/engine/screen_picker/plugin.hpp>
