+++ engine/src/physics/plugin.cpp
@@ -3,7 +3,7 @@
 #include <cubos/engine/collisions/plugin.hpp>
 #include <cubos/engine/fixed_step/plugin.hpp>
 #include <cubos/engine/physics/plugin.hpp>
-#include <cubos/engine/physics/solver/solver.hpp>
+#include <cubos/engine/physics/solver/plugin.hpp>
 #include <cubos/engine/settings/plugin.hpp>
 #include <cubos/engine/transform/plugin.hpp>
